Restructure in utils
